% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between_covariance.R
\docType{data}
\name{between_covariance}
\alias{between_covariance}
\title{Between source covariance matrix of eight trace copper wire chemical
concentrations.}
\format{
An 8 x 8 matrix.
}
\source{
\url{https://pubs.acs.org/doi/suppl/10.1021/ac5013514/suppl_file/ac5013514_si_002.docx}
}
\usage{
between_covariance
}
\description{
A covariance matrix providing the estimated variances and covariances of
the mean vector of eight chemical concentrations between multiple unique
sources. This covariance matrix is used for each component in a two component
Gaussian mixture model describing the mean chemical
concentrations of samples with different sources.
}
\keyword{datasets}
