% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_functions.R
\name{dtrace_term_dcov_par}
\alias{dtrace_term_dcov_par}
\title{Derivative of the trace term.}
\usage{
dtrace_term_dcov_par(cov_par, A_trace)
}
\arguments{
\item{cov_par}{A named list of covariance parameters.}

\item{A_trace}{A value calculated internally in the ELBO gradient function.}
}
\value{
The derivative of the trace term with respect to the nugget/noise.
}
\description{
Calculate the derivative of the trace term with respect to any covariance
parameters different from the nugget.
}
