% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_functions.R
\name{dtrace_term_dtau}
\alias{dtrace_term_dtau}
\title{Derivative of the trace term.}
\usage{
dtrace_term_dtau(cov_par, trace_term)
}
\arguments{
\item{cov_par}{A named list of covariance parameters.}

\item{trace_term}{Scalar value of the current trace term in the ELBO.}
}
\value{
The derivative of the trace term with respect to the nugget/noise.
}
\description{
Calculate the derivative of the trace term with respect to the nugget/noise.
}
