% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_cov_matC}
\alias{make_cov_matC}
\title{Make a covariance matrix with  covariance function}
\usage{
make_cov_matC(x, x_pred, cov_par, cov_fun, delta)
}
\arguments{
\item{x}{Matrix of observed data values}

\item{x_pred}{Potentially an empty matrix. If not empty, this returns
the covariance between x and x_pred.}

\item{cov_par}{Named list of covariance parameters. Need
"sigma", "tau", and "l1", ..., "ld" where d is the input dimension}

\item{cov_fun}{A character string specifying the covariance function.
Currently only works with "sqexp".}

\item{delta}{Scalar value that functions as a fixed nugget used to
stabilize matrix inverses and decompositions.}
}
\value{
Covariance matrix at the observed data locations if
x_pred = matrix(), or the covariances between x and x_pred if
x_pred is not an empty matrix.
}
\description{
Make a covariance matrix with  covariance function
}
