% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vi_functions.R
\name{trace_term_fun}
\alias{trace_term_fun}
\title{Calculate the trace term in the ELBO.}
\usage{
trace_term_fun(cov_par, Sigma12, Sigma22, delta)
}
\arguments{
\item{cov_par}{A named list of covariance parameters.}

\item{Sigma12}{Covariance matrix between observed data locations (rows) and knots (columns)}

\item{Sigma22}{Variance covariance matrix at the knots.}

\item{delta}{A small diagonal perturbation to Sigma22 for numerical stability.}
}
\value{
The value of the trace term in the ELBO.
}
\description{
Calculate the trace term in the ELBO for use in the variational approximation.
}
