% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within_covariance.R
\docType{data}
\name{within_covariance}
\alias{within_covariance}
\title{Within source covariance matrix of eight trace copper wire chemical
concentrations.}
\format{
An 8 x 8 matrix.
}
\source{
\url{https://pubs.acs.org/doi/suppl/10.1021/ac5013514/suppl_file/ac5013514_si_002.docx}
}
\usage{
within_covariance
}
\description{
A covariance matrix providing the estimated variances and covariances
from samples within a single source.
}
\keyword{datasets}
